+++ lib/syntax/lex.js
@@ -44,63 +44,3 @@ var Lex = (function (_super) {
     return Lex;
 }(frames_1.Frame));
 exports.Lex = Lex;
-var LexString = (function (_super) {
-    __extends(LexString, _super);
-    function LexString() {
-        return _super !== null && _super.apply(this, arguments) || this;
-    }
-    LexString.prototype.isEnd = function (char) {
-        return char === "”";
-    };
-    LexString.prototype.makeFrame = function () {
-        return new frames_1.FrameString(this.body);
-    };
-    return LexString;
-}(Lex));
-exports.LexString = LexString;
-;
-var LexComment = (function (_super) {
-    __extends(LexComment, _super);
-    function LexComment() {
-        return _super !== null && _super.apply(this, arguments) || this;
-    }
-    LexComment.prototype.isEnd = function (char) { return char === "#" || char === "\n"; };
-    LexComment.prototype.makeFrame = function () { return frames_1.FrameSymbol.for(""); };
-    return LexComment;
-}(Lex));
-exports.LexComment = LexComment;
-;
-var LexSpace = (function (_super) {
-    __extends(LexSpace, _super);
-    function LexSpace() {
-        return _super !== null && _super.apply(this, arguments) || this;
-    }
-    LexSpace.prototype.isEnd = function (char) { return char !== " "; };
-    LexSpace.prototype.makeFrame = function () { return frames_1.FrameSymbol.for(""); };
-    return LexSpace;
-}(Lex));
-exports.LexSpace = LexSpace;
-;
-var lex_routes = {
-    " ": new LexSpace(),
-    "#": new LexComment(),
-    "“": new LexString(),
-};
-var LexPipe = (function (_super) {
-    __extends(LexPipe, _super);
-    function LexPipe(out) {
-        var _this = this;
-        lex_routes[frames_1.Frame.kOUT] = out;
-        _this = _super.call(this, lex_routes) || this;
-        return _this;
-    }
-    LexPipe.prototype.lex_string = function (input) {
-        var source = new frames_1.FrameString(input);
-        return this.lex(source);
-    };
-    LexPipe.prototype.lex = function (source) {
-        return source.reduce(this);
-    };
-    return LexPipe;
-}(frames_1.Frame));
-exports.LexPipe = LexPipe;
