+++ lib/syntax/pipeline.d.ts
@@ -1,4 +1,9 @@
-import { Context, Frame } from "../frames";
+import { Context, Frame, FrameString } from "../frames";
+export declare class LexPipe extends Frame {
+    constructor(out: Frame);
+    lex_string(input: string): Frame;
+    lex(source: FrameString): Frame;
+}
 export declare class EvalPipe extends Frame {
     constructor(out: Frame, meta?: Context);
 }
