+++ lib/syntax/pipeline.js
@@ -5,7 +5,25 @@ var __extends = (this && this.__extends) || function (d, b) {
     d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
 };
 var frames_1 = require("../frames");
-var lex_1 = require("./lex");
+var tokens_1 = require("./tokens");
+var LexPipe = (function (_super) {
+    __extends(LexPipe, _super);
+    function LexPipe(out) {
+        var _this = this;
+        tokens_1.tokens[frames_1.Frame.kOUT] = out;
+        _this = _super.call(this, tokens_1.tokens) || this;
+        return _this;
+    }
+    LexPipe.prototype.lex_string = function (input) {
+        var source = new frames_1.FrameString(input);
+        return this.lex(source);
+    };
+    LexPipe.prototype.lex = function (source) {
+        return source.reduce(this);
+    };
+    return LexPipe;
+}(frames_1.Frame));
+exports.LexPipe = LexPipe;
 var EvalPipe = (function (_super) {
     __extends(EvalPipe, _super);
     function EvalPipe(out, meta) {
@@ -35,7 +53,7 @@ var piper = function (input, context) {
     var result = new frames_1.FrameArray([], context);
     var evaluator = new EvalPipe(result);
     var parser = new ParsePipe(evaluator);
-    var lexer = new lex_1.LexPipe(parser);
+    var lexer = new LexPipe(parser);
     var status = lexer.lex_string(input);
     if (status !== lexer) {
         console.error("\n* pipe returned " + status);
@@ -60,7 +78,7 @@ exports.framify_new = function (input, context) {
 };
 var pipe = function (input, out) {
     var output = new frames_1.FrameArray([]);
-    var lexer = new lex_1.LexPipe(output);
+    var lexer = new LexPipe(output);
     var status = lexer.lex_string(input);
     if (status !== lexer) {
         console.error("\n* pipe returned " + status);
