+++ lib/syntax/tokens.js
@@ -1,0 +1,50 @@
+"use strict";
+var __extends = (this && this.__extends) || function (d, b) {
+    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
+    function __() { this.constructor = d; }
+    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+};
+var frames_1 = require("../frames");
+var lex_1 = require("./lex");
+var LexString = (function (_super) {
+    __extends(LexString, _super);
+    function LexString() {
+        return _super !== null && _super.apply(this, arguments) || this;
+    }
+    LexString.prototype.isEnd = function (char) {
+        return char === "”";
+    };
+    LexString.prototype.makeFrame = function () {
+        return new frames_1.FrameString(this.body);
+    };
+    return LexString;
+}(lex_1.Lex));
+exports.LexString = LexString;
+;
+var LexComment = (function (_super) {
+    __extends(LexComment, _super);
+    function LexComment() {
+        return _super !== null && _super.apply(this, arguments) || this;
+    }
+    LexComment.prototype.isEnd = function (char) { return char === "#" || char === "\n"; };
+    LexComment.prototype.makeFrame = function () { return frames_1.FrameSymbol.for(""); };
+    return LexComment;
+}(lex_1.Lex));
+exports.LexComment = LexComment;
+;
+var LexSpace = (function (_super) {
+    __extends(LexSpace, _super);
+    function LexSpace() {
+        return _super !== null && _super.apply(this, arguments) || this;
+    }
+    LexSpace.prototype.isEnd = function (char) { return char !== " "; };
+    LexSpace.prototype.makeFrame = function () { return frames_1.FrameSymbol.for(""); };
+    return LexSpace;
+}(lex_1.Lex));
+exports.LexSpace = LexSpace;
+;
+exports.tokens = {
+    " ": new LexSpace(),
+    "#": new LexComment(),
+    "“": new LexString(),
+};
