+++ src/syntax/lex.ts
@@ -41,47 +41,3 @@ export class Lex extends Frame {
     return Frame.nil;
   }
 }
-
-export class LexString extends Lex {
-  protected isEnd(char: string) {
-    return char === "”";
-  }
-
-  protected makeFrame() {
-    return new FrameString(this.body);
-  }
-};
-
-export class LexComment extends Lex {
-  protected isEnd(char: string) { return char === "#" || char === "\n"; }
-
-  protected makeFrame() { return FrameSymbol.for(""); }
-};
-
-export class LexSpace extends Lex {
-  protected isEnd(char: string) { return char !== " "; }
-
-  protected makeFrame() { return FrameSymbol.for(""); }
-};
-
-const lex_routes: Context = {
-  " ": new LexSpace(),
-  "#": new LexComment(),
-  "“": new LexString(),
-};
-
-export class LexPipe extends Frame {
-  constructor(out: Frame) {
-    lex_routes[Frame.kOUT] = out;
-    super(lex_routes);
-  }
-
-  public lex_string(input: string) {
-    const source = new FrameString(input);
-    return this.lex(source);
-  }
-
-  public lex(source: FrameString) {
-    return source.reduce(this);
-  }
-}
